requirements.txt
Django==4.2.7
django-allauth==0.57.0